Fully commented but live audio doesnt work still

+++ InputClasses/InputDecorator.pde
@@ -1,21 +1,30 @@
 /**
-* The Input Decorator is an abstract interface that contains
-* reference to the decorated object (MIDI input).  Specific decorators
-* like SoundInput add additional decorating behavior.
+* The Input Decorator is an abstract class that contains a
+* reference to the base input object.  Specific decorators
+* like SoundDecorator add additional decorating behavior.
 */
-
 class InputDecorator extends input {
 
+    //Reference to base input object
     public input inputComponent;
 
-    public float decorateVal;
+    //Value used to decorate 
+    public double decorateVal;
     
+    /**
+    * Constructor initializes reference to base input
+    * @param inputCom : reference to base input object
+    */
     InputDecorator(input inputComp) {   
       inputComponent = inputComp;
       decorateVal = 0; 
     }
     
-    public float getVal(){
+    /**
+    * Returns value, overloads super class method
+    * @return sum of base value and decorate value
+    */
+    public double getVal(){
       return (inputComponent.getVal()+decorateVal);
     }
